+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.1] - Unreleased
+### Fixed
+- TypeError on restart the server.
+- Inherit stdout and stderr if `--show-terminal` is disabled.
+
 ## [0.2.0] - 2025-03-16
 ### Added
 - Improved cold start with more feedback.
@@ -24,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2025-01-16
 First version
 
+[0.2.1]: https://github.com/oscarotero/cms-lume-adapter/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/oscarotero/cms-lume-adapter/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/oscarotero/cms-lume-adapter/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/oscarotero/cms-lume-adapter/compare/v0.1.1...v0.1.2